customize: start page as custom page/route
== Setting / scope

This page is meant as entrypoint for the public IONOS GPT demo page
before the user has a session.

== Navigation

goto() was deliberrately avoided and window.location to bypass the
router and just start with a fresh page.

== Expectation

The page was loaded with a Cookie containing a session ID that the
GPT demo management backend can use to start a Open WebUI session.

== Limitation

Default model still hardcoded. According to ##151 this needs to be
provided by a config or as suggestion by a (virtual) model info.