customize: stop full init in root layout when at /start
The route /start is meant to contain the landing page of the GPT demo.

Avoid dynamically initializing all application and Websockets at this
point as we want to start the application after the user picked a choice
here.

We might later have to move this landing page out of Open Web UI.