customize(Chat): change chat prompt init from ?q= to sessionStorage
The sessionStorage entry "ionosGptDemoPrompt" is introduced. Another
part of the application can set a prompt to this entry and that chat
will load the prompt form this entry.

This avoid passing possibly sensitive prompts via query parameter, when
loading the chat with a hard page load, which would fetch the resource
from the server.