coustomize(router): do not redirect from /auth to /start
While integrating with the management service we discovered that the
code has to go via /auth to trigger a (transparent) sign-on.

We'll allow this by removing this config.

From the start page we instead go to /auth now to trigger the login
first.

The preselected prompt and model should be retained in the
sessionStorage.

Since this can not easily tested locally without the management service,
we'll deploy and test this on the staging system.