[POC] customize(Chat): goto /start not / for new chat button
Pragmatic change.

Three problems with the existing spaghetti code:

1. The Svelte code should not interact with DOM elements across
   components
2. The routing is implemented in at least two places
3. The explicit goto('/') was no-op since the href had already changed
   the location (thus adding the preventDefault modifier)

== Limitation

Directly navigating to / will still render Chat

For that we probably have to either change every place that explicitly
routes to / or, better, change the approach and change / to render our
own start page. This would imply bigger changes to the routing in OWU
and to our management service too.

Refs: ##205