c/feat: add startup module
== Context

Agent and optionally prompty have been saved to localStorage on the
startpage (explore) as preselection.

Since this page is also reached by unauthenticated users, we can not
use all APIs to store the agent (model) and init chat.

== Requirement

A user should be able to select and agent (model) and/or prompt (implies
an agent) on the startpage (explore) and jump right into the empty chat
or stated conversation.

This should work for both authenticated users and users that will go
though the (possibly) external authentication.

== Solution

The module loads the stored agent and prompt from localStorage and
persists the agent in the settings.

At least the model (agent) or both are returned for consumption by the
chat to start a conversation.

Refs: PRODAI-105