[DEBUG] Add debug logging to trace possible race condition
== Symptom

Sometimes chat does not start and stays in loading condition without
receiving and rendering messages when chat was started from a prompt
proposal on the Startpage.

This is obviously a race condition, but it depends on some initial
condition, and the steps to reproduce for these condidtions are not
known yet.

STR:

1. Start chat from Startpage prompt proposal
2. Wait for chat to complete
3. Go to Startpage (via Explore IONOS GPT)
4. Start chat from Startpage prompt proposal

Expected: chat is always started
Actual: sometimes it's started, sometimes it fails

== Original error

```
Chat.svelte:1406 Uncaught (in promise) TypeError: Cannot read properties of undefined (reading 'files')
    at Chat.svelte:1406:15
    at Array.some (<anonymous>)
    at Chat.svelte:1405:33
    at Array.map (<anonymous>)
    at sendPrompt (Chat.svelte:1398:21)
    at async submitPrompt (Chat.svelte:1327:14)

Messages.svelte:378
```

== Console log

```
 ## initNewChat(): reset history
 ## submitPrompt: add message to history with ID b01483a1-5eed-4d95-969f-07b3707ed1b8
 ## initNewChat(): reset history
 Message not found in history by ID  b01483a1-5eed-4d95-969f-07b3707ed1b8
 <repeated multiple times>
 ## sendPrompt: messages= [] for parentId= b01483a1-5eed-4d95-969f-07b3707ed1b8
```